backend/pyproject.toml
[project]
name = "crm-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.0",
  "pydantic[email]>=2.10.0",
  "redis[hiredis]>=5.2.0",
  "bcrypt>=4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.12",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
